update okhttp version, made inner class CustomWebSocketListener
