+++ src/main/kotlin/playwright/websocket/provider/WebSocketProvider.kt
@@ -1,35 +1,49 @@
 package playwright.websocket.provider
 
 import core.exceptions.WebSocketException
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.WebSocket
-import playwright.websocket.listener.WebSocketListener
+import okhttp3.*
 import java.time.Duration
-
+import java.util.concurrent.BlockingQueue
+import java.util.concurrent.LinkedBlockingQueue
 import java.util.concurrent.TimeUnit
 
-class WebSocketProvider(_webSocketListener: WebSocketListener) {
-    private val client: OkHttpClient
-    private val webSocketListener: WebSocketListener
-    private val defaultTimeOut: Long = 3
-    private lateinit var webSocket: WebSocket
-    private val normalClosureStatus: Int = 1000
+class WebSocketProvider {
+    private class CustomWebSocketListener : WebSocketListener() {
+        private val incomingMessages: BlockingQueue<String> = LinkedBlockingQueue()
 
-    init {
-        this.client = OkHttpClient.Builder()
-            .readTimeout(defaultTimeOut, TimeUnit.SECONDS)
-            .build()
-        this.webSocketListener = _webSocketListener
+        override fun onFailure(webSocket: WebSocket, t: Throwable, response: Response?) {
+            super.onFailure(webSocket, t, response)
+            throw WebSocketException("Websocket has been closed due to an error reading from or writing to the network!", t)
+        }
+
+
+        override fun onMessage(webSocket: WebSocket, text: String) {
+            super.onMessage(webSocket, text)
+            incomingMessages.add(text)
+        }
+
+        fun pollMessage(timeout: Long, timeUnit: TimeUnit): String? = incomingMessages.poll(timeout, timeUnit)
     }
 
+    private val defaultTimeOut: Long = 3
+    private lateinit var client: OkHttpClient
+    private lateinit var webSocket: WebSocket
+    private val normalClosureStatus: Int = 1000
+    private lateinit var webSocketListener: CustomWebSocketListener
+
     fun connect(url: String) {
+        client = OkHttpClient.Builder()
+            .readTimeout(defaultTimeOut, TimeUnit.SECONDS)
+            .build()
+
+        webSocketListener = CustomWebSocketListener()
+
         val request: Request = Request.Builder()
             .url(url)
             .build()
 
         try {
-            webSocket = client.newWebSocket(request, this.webSocketListener)
+            webSocket = client.newWebSocket(request, webSocketListener)
         } catch (e: Exception) {
             throw WebSocketException("Connection failed!", e.cause)
         }
@@ -39,7 +53,8 @@ class WebSocketProvider(_webSocketListener: WebSocketListener) {
         this.webSocket.send(message)
     }
 
-    fun pollMessage(timeout: Duration) : String? = webSocketListener.pollMessage(timeout.toMillis(), TimeUnit.MILLISECONDS)
+    fun pollMessage(timeout: Duration): String? =
+        webSocketListener.pollMessage(timeout.toMillis(), TimeUnit.MILLISECONDS)
 
     fun closeConnection() {
         this.webSocket.close(normalClosureStatus, "Close connection")
